# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mi Negocio Pro API",
        "description": "API para gestión completa de negocio: usuarios, productos, ventas, compras y más",
        "contact": {
            "email": "support@minegocio.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server Local"
        }
    ],
    "paths": {
...
        "/api/purchase-items": {
            "get": {
                "tags": ["Purchase Items"],
                "summary": "Listar items de compras",
                "description": "Obtiene una lista paginada de items de compras con filtros opcionales",
                "operationId": "f1ab0227c0511c7ac5c48a694450d7b5",
                "parameters": [
                    {
                        "name": "purchase_id",
                        "in": "query",
                        "description": "Filtrar por ID de compra",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "product_id",
                        "in": "query",
                        "description": "Filtrar por ID de producto",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "min_quantity",
                        "in": "query",
                        "description": "Cantidad mínima",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "max_unit_cost",
                        "in": "query",
                        "description": "Costo unitario máximo",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "with_relations",
                        "in": "query",
                        "description": "Incluir relaciones (producto y compra)",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de items de compras",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/PurchaseItems"
                                            }
                                        },
                                        "current_page": {
                                            "type": "integer"
                                        },
                                        "last_page": {
                                            "type": "integer"
                                        },
                                        "per_page": {
                                            "type": "integer"
                                        },
                                        "total": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": ["Purchase Items"],
                "summary": "Crear item de compra",
                "description": "Crea un nuevo item de compra",
                "operationId": "cf2b653075012b51604daaf270132155",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "purchase_id",
                                    "product_id",
                                    "quantity",
                                    "unit_cost"
                                ],
                                "properties": {
                                    "purchase_id": {
                                        "description": "ID de la compra",
                                        "type": "integer"
                                    },
                                    "product_id": {
                                        "description": "ID del producto",
                                        "type": "integer"
                                    },
                                    "quantity": {
                                        "description": "Cantidad",
                                        "type": "integer"
                                    },
                                    "unit_cost": {
                                        "description": "Costo unitario",
                                        "type": "number",
                                        "format": "float"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Item de compra creado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Item de compra creado exitosamente"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/PurchaseItems"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/purchase-items/{id}": {
            "get": {
                "tags": ["Purchase Items"],
                "summary": "Obtener item de compra",
                "description": "Obtiene los detalles de un item de compra específico",
                "operationId": "7a85221db21a9b31e69a08732527743f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del item de compra",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del item de compra",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PurchaseItems"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Item de compra no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": ["Purchase Items"],
                "summary": "Actualizar item de compra",
                "description": "Actualiza los datos de un item de compra existente",
                "operationId": "1c5b2896ff853867b765ea142f089509",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del item de compra",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "purchase_id": {
                                        "description": "ID de la compra",
                                        "type": "integer"
                                    },
                                    "product_id": {
                                        "description": "ID del producto",
                                        "type": "integer"
                                    },
                                    "quantity": {
                                        "description": "Cantidad",
                                        "type": "integer"
                                    },
                                    "unit_cost": {
                                        "description": "Costo unitario",
                                        "type": "number",
                                        "format": "float"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Item de compra actualizado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Item de compra actualizado exitosamente"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/PurchaseItems"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": ["Purchase Items"],
                "summary": "Eliminar item de compra",
                "description": "Elimina un item de compra existente",
                "operationId": "76775767b76154cfd63e88c7cc01a358",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del item de compra",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item de compra eliminado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Item de compra eliminado exitosamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/purchases/{purchase_id}/items": {
            "get": {
                "tags": ["Purchase Items"],
                "summary": "Obtener items de una compra",
                "description": "Obtiene todos los items de una compra específica",
                "operationId": "a978377e19694d12e43c952944d27fce",
                "parameters": [
                    {
                        "name": "purchase_id",
                        "in": "path",
                        "description": "ID de la compra",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Items de la compra",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PurchaseItems"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/purchase-items/stats": {
            "get": {
                "tags": ["Purchase Items"],
                "summary": "Estadísticas de items de compras",
                "description": "Obtiene estadísticas generales de los items de compras",
                "operationId": "af4cf96ef37ad88a3fba82e1fd027d12",
                "responses": {
                    "200": {
                        "description": "Estadísticas de items de compras",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "general_stats": {
                                            "type": "object"
                                        },
                                        "most_purchased_products": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "recent_purchases": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "cost_analysis": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "top_suppliers": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/products/{product_id}/purchase-history": {
            "get": {
                "tags": ["Purchase Items"],
                "summary": "Historial de compras de un producto",
                "description": "Obtiene el historial de compras de un producto específico",
                "operationId": "2eb99de4c04971126465d32a319ca166",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path",
                        "description": "ID del producto",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Número máximo de registros",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Historial de compras del producto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PurchaseItems"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}